Factor out indexedSend helper function into utilities module
